examples/latin_lineation_mvp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "latin-lineation-mvp"
version = "0.1.0"
description = "Train per-line masks from latin_documents PageXML; predict_masks for transcriber-shell"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26.0",
  "pillow>=10.0.0",
  "torch>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
latin-lineation-train = "latin_lineation_mvp.train:main"
latin-lineation-benchmark = "latin_lineation_mvp.benchmark:main"
